add ByteType (derived from IntegerType), and use this for protocol fields (SNMP version, error status / non-repeaters, error index / repeaters). add _knownLen to IntegerType: defaults to 4 (1 for ByteType), and is set to the decoded _length in fromBuffer(). this is then used in serialise(), where it selectively outputs extra bytes only when needed.
with all that, the resultant outgoing packet will be at least 9 bytes shorter.

for certain packets, it could be 10 bytes shorter, if the decrease by 9 forces a containing struct's length to drop from requiring 2 bytes to encode its length down to 1, etc.

that is the case for the test packet returned by the test.cpp GenerateTestSNMPRequestPacket() function.
it was generating a 133 byte packet, but now is only 123.

update the unit tests to now use a const (const int expected_length = 123;) throughout.

+++ src/include/SNMPPacket.h
@@ -73,7 +73,7 @@ class SNMPPacket {
     bool reuse = false;
 
     std::shared_ptr<IntegerType> requestIDPtr = nullptr;
-    std::shared_ptr<IntegerType> snmpVersionPtr = nullptr;
+    std::shared_ptr<ByteType> snmpVersionPtr = nullptr;
     std::shared_ptr<OctetType> communityStringPtr = nullptr;
 
     snmp_request_id_t requestID = 0;
